waste_classifier_service/requirements.txt
Flask==2.3.3
gunicorn==21.2.0
Pillow==10.0.1
tensorflow==2.12.0
scikit-learn==1.3.0
numpy==1.23.5  # <-- The changed line
requests==2.31.0
python-dotenv==1.0.0